# docs/swagger.yaml
basePath: /api/v1
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  jsontype.Tags:
    additionalProperties: {}
    type: object
...
  omc.Device:
    properties:
      ActiveStatus:
        type: string
      Created:
        type: string
      Deleted:
        $ref: '#/definitions/gorm.DeletedAt'
      Enable:
        type: boolean
      Group:
        $ref: '#/definitions/omc.Group'
      GroupId:
        type: integer
      Id:
        type: integer
      LastInformTime:
        type: string
      MetaData:
        $ref: '#/definitions/jsontype.Tags'
      Methods:
        items:
          type: string
        type: array
      Name:
        type: string
      Online:
        type: boolean
      Oui:
        type: string
      ParameterNotifications:
        $ref: '#/definitions/omc.ParameterNotifications'
      ParameterValues:
        $ref: '#/definitions/omc.ParameterValues'
      ParameterWritables:
        $ref: '#/definitions/omc.ParameterWritables'
      Product:
        $ref: '#/definitions/omc.Product'
      ProductClass:
        type: string
      ProductId:
        type: integer
      ProductType:
        type: string
      Properties:
        $ref: '#/definitions/jsontype.Tags'
      Schema:
        type: string
      SerialNumber:
        type: string
      Updated:
        type: string
    type: object
...
  omc.Firmware:
    properties:
      Created:
        type: string
      Deleted:
        $ref: '#/definitions/gorm.DeletedAt'
      Id:
        type: integer
      Name:
        type: string
      ProductType:
        type: string
      S3Object:
        $ref: '#/definitions/s3.S3Object'
      S3ObjectId:
        type: integer
      Schema:
        type: string
      Updated:
        type: string
      UploadTime:
        type: string
      Uploader:
        type: string
      Version:
        type: string
      products:
        items:
          $ref: '#/definitions/omc.Product'
        type: array
    type: object
  omc.Group:
    properties:
      Children:
        items:
          $ref: '#/definitions/omc.Group'
        type: array
      Created:
        type: string
      Id:
        type: integer
      MetaData:
        $ref: '#/definitions/jsontype.Tags'
      Name:
        type: string
      Parent:
        $ref: '#/definitions/omc.Group'
      ParentId:
        type: integer
      Schema:
        type: string
      Updated:
        type: string
    type: object
  omc.KPIMeas:
    properties:
      Created:
        type: string
      Default:
        type: boolean
      Deleted:
        $ref: '#/definitions/gorm.DeletedAt'
      Enable:
        type: boolean
      Formula:
        type: string
      Id:
        type: integer
      MeasTypeID:
        type: string
      MeasTypeSet:
        type: string
      MetaData:
        $ref: '#/definitions/jsontype.Tags'
      Name:
        type: string
      ProductType:
        type: string
      Schema:
        type: string
      StatsType:
        type: string
      Unit:
        type: string
      Updated:
        type: string
    type: object
...
  omc.ParameterNotifications:
    additionalProperties:
      type: integer
    type: object
  omc.ParameterValues:
    additionalProperties:
      type: string
    type: object
  omc.ParameterWritables:
    additionalProperties:
      type: boolean
    type: object
  omc.Product:
    properties:
      Created:
        type: string
      Default:
        type: boolean
      Deleted:
        $ref: '#/definitions/gorm.DeletedAt'
      Enable:
        type: boolean
      Firmwares:
        items:
          $ref: '#/definitions/omc.Firmware'
        type: array
      Id:
        type: integer
      Manufacturer:
        type: string
      MetaData:
        $ref: '#/definitions/jsontype.Tags'
      ModelName:
        type: string
      Oui:
        type: string
      ParameterPath:
        type: string
      PerformanceValueDefines:
        $ref: '#/definitions/jsontype.Tags'
      ProductClass:
        type: string
      ProductType:
        type: string
      Schema:
        type: string
      SupportedAlarms:
        items:
          $ref: '#/definitions/omc.ProductSupportedAlarm'
        type: array
      Updated:
        type: string
    type: object
  omc.ProductSupportedAlarm:
    properties:
      AlarmIdentifier:
        type: string
      Created:
        type: string
      EventType:
        type: string
      Id:
        type: integer
      Oui:
        type: string
      PerceivedSeverity:
        type: string
      ProbableCause:
        type: string
      Product:
        $ref: '#/definitions/omc.Product'
      ProductClass:
        type: string
      ProductId:
        type: integer
      ProductType:
        type: string
      Schema:
        type: string
      SerialNumber:
        type: string
      SpecificProblem:
        type: string
      Updated:
        type: string
    type: object
  omc.Task:
    properties:
      Created:
        type: string
      Creater:
        type: string
      Default:
        type: boolean
      Deleted:
        $ref: '#/definitions/gorm.DeletedAt'
      DeviceIds:
        $ref: '#/definitions/jsontype.Tags'
      Enable:
        type: boolean
      ExecEndTime:
        type: string
      ExecInterval:
        type: integer
      ExecLastTime:
        type: string
      ExecMode:
        $ref: '#/definitions/omc.TaskExecMode'
      ExecNextTime:
        type: string
      ExecProcess:
        type: integer
      ExecRate:
        type: integer
      ExecStartTime:
        type: string
      ExecTimes:
        type: integer
      Id:
        type: integer
      MetaData:
        $ref: '#/definitions/jsontype.Tags'
      RetryInterval:
        type: integer
      RetryTimes:
        type: integer
      Schema:
        type: string
      TaskName:
        type: string
      TaskStatus:
        $ref: '#/definitions/omc.TaskStatus'
      TaskType:
        type: string
      Updated:
        type: string
    type: object
  omc.TaskDeviceLog:
    properties:
      Code:
        type: integer
      Device:
        $ref: '#/definitions/omc.Device'
      DeviceId:
        type: integer
      DeviceIds:
        $ref: '#/definitions/jsontype.Tags'
      EndTime:
        type: string
      ExecMode:
        type: integer
      Info:
        type: string
      Schema:
        type: string
      StartTime:
        type: string
      Task:
        $ref: '#/definitions/omc.Task'
      TaskId:
        type: integer
      TaskName:
        type: string
      TaskType:
        type: string
      Time:
        type: string
      Updated:
        type: string
    type: object
  omc.TaskExecMode:
    enum:
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - TaskExecModeImmediately
    - TaskExecModeSchedule
  omc.TaskStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - TaskStatusInit
    - TaskStatusWaiting
    - TaskStatusRunning
    - TaskStatusFinish
    - TaskStatusCancel
...
  omc.createKPIMeasureBody:
    properties:
      formula:
        type: string
      measTypeID:
        type: string
      measTypeSet:
        type: string
      name:
        type: string
      productType:
        type: string
      statsType:
        type: string
      unit:
        type: string
    type: object
...
  omc.listKPIMeasuresData:
    properties:
      Data:
        items:
          $ref: '#/definitions/omc.KPIMeas'
        type: array
      Total:
        type: integer
    type: object
...
  omc.listTaskDeviceLogsData:
    properties:
      Data:
        items:
          $ref: '#/definitions/omc.TaskDeviceLog'
        type: array
      Total:
        type: integer
    type: object
...
  omc.updateDeviceInfoBody:
    properties:
      Name:
        type: string
    type: object
...
  omc.updateKPIMeasureInfoBody:
    properties:
      name:
        type: string
    type: object
...
  s3.S3Object:
    properties:
      Bucket:
        type: string
      Created:
        type: string
      Deleted:
        $ref: '#/definitions/gorm.DeletedAt'
      DownloadUrl:
        type: string
      FileName:
        type: string
      FileSize:
        type: integer
      Id:
        type: integer
      Key:
        type: string
      MetaData:
        $ref: '#/definitions/s3.Tags'
      Schema:
        type: string
      SignMethod:
        type: string
      Signature:
        type: string
      Updated:
        type: string
    type: object
  s3.Tags:
    additionalProperties: {}
    type: object
...
  system.systemInfoData:
    properties:
      Build:
        type: string
      Name:
        type: string
      Version:
        type: string
    type: object
  system.systemTimeData:
    properties:
      Current:
        type: integer
    type: object
...
host: dev.netdoop.com
info:
  contact: {}
  title: NetDoop API
  version: "1.0"
paths:
...
  /omc/device-logs:
    get:
      operationId: list-device-logs
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - default: ""
        description: Sort order
        in: query
        name: order_by
        type: string
      - default: ""
        description: Query
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total:
              description: Total number
              type: int
          schema:
            $ref: '#/definitions/omc.listTaskDeviceLogsData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: List log device logs
      tags:
      - OMC TaskDeviceLogs
  /omc/device-logs/{ts}:
    get:
      operationId: get-device-log
      parameters:
      - description: Timestamp
        in: path
        name: ts
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/omc.TaskDeviceLog'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Get log device log
      tags:
      - OMC TaskDeviceLogs
...
  /omc/devices/{id}:
    delete:
      consumes:
      - application/json
      operationId: delete-device
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Delete a device
      tags:
      - OMC Devices
    get:
      consumes:
      - application/json
      operationId: get-device
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/omc.Device'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Get a device
      tags:
      - OMC Devices
    put:
      consumes:
      - application/json
      operationId: update-device-info
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: integer
      - description: Device
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/omc.updateDeviceInfoBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/omc.Device'
        "400":
          description: 'Bad Request: invalid input parameter'
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Update device info
      tags:
      - OMC Devices
...
  /omc/kpi/measures:
    get:
      operationId: list-kpi-measures
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - default: ""
        description: Sort order
        in: query
        name: order_by
        type: string
      - default: ""
        description: Query
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total:
              description: Total number
              type: int
          schema:
            $ref: '#/definitions/omc.listKPIMeasuresData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: List kpi measures
      tags:
      - OMC KPI
    post:
      operationId: create-kpi-measure
      parameters:
      - description: KPIMeas
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/omc.createKPIMeasureBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/omc.KPIMeas'
        "400":
          description: 'Bad Request: invalid input parameter'
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Create kpi measure
      tags:
      - OMC KPI
  /omc/kpi/measures/{id}:
    delete:
      operationId: delete-kpi-measure
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Delete kpi measure
      tags:
      - OMC KPI
    get:
      operationId: get-kpi-measure
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/omc.KPIMeas'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Get kpi measure
      tags:
      - OMC KPI
    put:
      operationId: update-kpi-measure-info
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: KPIMeas
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/omc.updateKPIMeasureInfoBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/omc.KPIMeas'
        "400":
          description: 'Bad Request: invalid input parameter'
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Update kpi measure info
      tags:
      - OMC KPI
  /omc/kpi/measures/{id}/disable:
    put:
      operationId: set-kpi-measure-disable
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/omc.KPIMeas'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Set kpi measure disable
      tags:
      - OMC KPI
  /omc/kpi/measures/{id}/enable:
    put:
      operationId: set-kpi-measure-enable
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/omc.KPIMeas'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Set kpi measure enable
      tags:
      - OMC KPI
...
  /system/info:
    get:
      operationId: get-system-info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/system.systemInfoData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Get information of system
      tags:
      - System
  /system/time:
    get:
      operationId: get-system-time
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/system.systemTimeData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Get current time of system
      tags:
      - System
schemes:
- http
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"